Rename projects
Preparations for the java transformation project.

All projects will use the 'tools.vitruv.applications.asemsysml' prefix.
After that, the project specific purpose is used for the project name.
E.g.'.java' for the java transformations or '.tests' for the test cases.
At the end the transformation direction is specified. In this case
'.sysml2asem'.