Add tests and transformations for SysML port transformation
The java and reactions transformation will transform the port of a SysML
block to a ASEM message of a ASEM module. The names of the port and the
message will be equal and the port direction is mapped to the messages
readable and writable attributes.

The port type will be not transformed at the moment!
Nested ports are ignored, too.