Add port type transformation
A port which is typed by a SysML block will be transformed to a ASEM
message. The message will be typed by an ASEM module, if the block which
types the port corresponds to an ASEM module, or by an ASEM class, if
this block corresponds to an ASEM class.

Value types as port types are not handled at the moment.